tidy up tests and move to blogTest
